Arreglado los errores de mismas favoritas para distintos usuarios y eliminacion de peliculas en el top 10

+++ app/src/main/java/es/riberadeltajo/ceca_guillermoimdbapp/database/FavoritesDatabaseHelper.java
@@ -17,6 +17,7 @@ public class FavoritesDatabaseHelper extends SQLiteOpenHelper {
     public static final String COLUMN_TITLE = "title";
     public static final String COLUMN_RELEASE_DATE = "release_date";
     public static final String COLUMN_POSTER_PATH = "poster_path";
+    public static final String COLUMN_USER_ID = "user_id";
 
     public FavoritesDatabaseHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -26,17 +27,23 @@ public class FavoritesDatabaseHelper extends SQLiteOpenHelper {
     @Override
     public void onCreate(SQLiteDatabase db) {
         String CREATE_FAVORITES_TABLE = "CREATE TABLE " + TABLE_FAVORITES + " (" +
-                COLUMN_ID + " TEXT PRIMARY KEY, " +
+                COLUMN_ID + " TEXT, " +
                 COLUMN_TITLE + " TEXT, " +
                 COLUMN_RELEASE_DATE + " TEXT, " +
                 COLUMN_RATING + " TEXT, " +
-                COLUMN_POSTER_PATH + " TEXT)";
+                COLUMN_POSTER_PATH + " TEXT, " +
+                COLUMN_USER_ID + " TEXT, " +
+                "PRIMARY KEY (" + COLUMN_ID + ", " + COLUMN_USER_ID + "))";  // Clave primaria combinada
         db.execSQL(CREATE_FAVORITES_TABLE);
     }
 
+
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_FAVORITES);
-        onCreate(db);
+        if (oldVersion < 3) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_FAVORITES);
+            onCreate(db);
+        }
     }
+
 }
